Adding path to find route

+++ BeagleDarc/GUI/BeagleDarcGui.py
@@ -1,7 +1,11 @@
+
+import os
 import sys
 import pygtk  
 pygtk.require("2.0")  
 import gtk  
+
+from subprocess import Popen, PIPE
 
 from BeagleDarc.Model import Model
 
@@ -10,6 +14,7 @@ class BeagleDarcGui:
     wTree = gtk.Builder()
 
     def __init__( self ):
+        path, fil = os.path.split(os.path.abspath(__file__))
         self.model = Model()
         self.serverhost = self.model.get_beagledarc_server_host()
         self.serveruser = self.model.get_beagledarc_server_user()
@@ -17,13 +22,18 @@ class BeagleDarcGui:
         self.serverport = self.model.get_beagledarc_server_port()
 
         self.builder = gtk.Builder()
-        self.builder.add_from_file("glade/beagledarc.glade")
+        self.builder.add_from_file(path+"/glade/beagledarc.glade")
         self.window = self.builder.get_object ("window1")
-#        self.window.set_icon_from_file('img/icon-beagle.png')
+        self.window.set_events(self.window.get_events() | gtk.gdk.BUTTON_PRESS_MASK)
 
         if self.window:
             self.window.connect("destroy", gtk.main_quit)
 
+        #Toggle button to connect to beaglebone
+        self.connect_togglebutton = self.builder.get_object ("connect_togglebutton")
+        self.connect_togglebutton.connect("toggled", self.callback, "Connection")
+
+        #default entries
         self.entry1 = self.builder.get_object ("entry1")
         self.entry2 = self.builder.get_object ("entry2")
         self.entry3 = self.builder.get_object ("entry3")
@@ -39,22 +49,30 @@ class BeagleDarcGui:
             "on_window1_destroy" : self.quit,
             "gtk_widget_destroy" : self.quit,
             "on_phasescreen_menuitem_activate" : self.phasescreen,
-            "on_stars_menuitem_activate" : self.stars,
-            "on_connect_togglebutton_toggled": self.connect
+            "on_stars_menuitem_activate" : self.stars
         }
         
         self.builder.connect_signals( dic )
 
+    def callback(self, widget, data=None):
+        '''
+        callback
+        '''
+        print "%s was toggled %s" % (data, ("OFF", "ON")[widget.get_active()])
+        if widget.get_active() is True:
+            cmd = "ssh %s@%s \"python /home/root/server.py &\"" % (self.serveruser, self.serverhost)
+        if widget.get_active() is False:
+            cmd = "ssh %s@%s \"ps aux |grep server.py|awk \'{print \\$2}\'|xargs kill -9\"" % (self.serveruser, self.serverhost)
+        #process = Popen(cmd , stdout=PIPE , stderr=PIPE , shell=True)
+        process = Popen(cmd , stdout=sys.stdout , stderr=sys.stderr , shell=True)
+        sts = process.wait()
+        #out = process.stdout.read().strip()
+        #err = process.stderr.read().strip()
+        #print cmd
+
     def quit(self, widget):
         sys.exit(0)
 
-    def connect(self, widget):
-        print "connected :)"
-        self.disconnect(widget)
-
-    def disconnect(self, widget):
-        cmd = "ssh %s@%s \"ps aux |grep server.py|awk \'{print \\$2}\'|xargs kill -9\"" % (self.serveruser, self.serverhost)
-        print cmd
 
     def phasescreen(self, widget):
         print "phasescreen"
